refactor(system-agent): split passive QR lifecycle
